LPS-66791 Update the tests
